Refactor: Move logic into internal module
Moved several internal functions such as `read_patient_id_from_file` and `dcm_cp_file` into an `internal` module for better encapsulation and organization. Also, simplified the main function by using the new `dcm_cp_files` function.
